datasets/requirements.txt
salesforce-lavis
scenedetect[opencv]
boto3
moviepy
git+https://github.com/huggingface/transformers
